requirements.txt
bokeh==1.3.4
Django==2.2.4
django-extensions==2.2.1
djangorestframework==3.10.2
geopy==1.20.0
matplotlib==3.1.1
networkx==2.4
numpy==1.17.0
Pillow==6.1.0
requests==2.22.0